pyproject.toml
[tool.poetry]
name = "CalibreOneDriveSync"
version = "0.1.0"
description = "A Python project to synchronize Calibre library with OneDrive and reload CalibreWeb DB when changes are detected"
authors = ["ENDMVP"]
license = "MIT"

[tool.poetry.dependencies]
requests = "^2.31.0"
python = "^3.8"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
